Fix Precompile Address Width (#396)
When looking through the code, I realized that I made a mistake in #379.
Specifically, we specify that precompiles fit into a `uint32`, when in
fact they fit into a `uint16`:

https://eips.ethereum.org/EIPS/eip-1352

This PR changes the code to encode the `P256.Verifiers` type to be a
`uint16 || address` instead of a slightly larger `uint32`.

Alternatively, we _could_ use a full 96 bits to represent the precompile
address and make `P256.Verifiers` a `uint256` (this has a small positive
effect on gas usage and code size, as it gets rid of compiler generated
masking code).

+++ examples/4337-passkeys/src/logic/safe.ts
@@ -61,7 +61,7 @@ function getInitHash(safeInitializer: SafeInitializer, chainId: ethers.BigNumber
         { type: 'address', name: 'signerFactory' },
         { type: 'uint256', name: 'signerX' },
         { type: 'uint256', name: 'signerY' },
-        { type: 'uint192', name: 'signerVerifiers' },
+        { type: 'uint176', name: 'signerVerifiers' },
         { type: 'address', name: 'setupTo' },
         { type: 'bytes', name: 'setupData' },
         { type: 'address', name: 'fallbackHandler' },
